+++ zqd/storage/storage.go
@@ -1,0 +1,9 @@
+package storage
+
+import (
+	"github.com/brimsec/zq/zqd/storage/unizng"
+)
+
+func Load(path string) (*unizng.ZngStorage, error) {
+	return unizng.Load(path)
+}
