+++ zqd/storage/unizng/driver.go
@@ -1,4 +1,4 @@
-package storage
+package unizng
 
 import (
 	"github.com/brimsec/zq/zbuf"
